requirements.txt
Django>=4.0.4,<4.1
djangorestframework>=3.13.1,<3.14
mysqlclient
pymysql
pillow
djangorestframework-simplejwt
google-cloud-vision
drf-yasg
python-dotenv
requests
Pillow  # For ImageField
celery